test(localio): isolate input boundary e2e
